Implemented auto-loading of templates.
It is not working, so its left in this branch for further investigation.
The issue is:
Having the templates in <script type=xxx>template A<script> does not
let us reuse the templates for other purposes. (i.e testing in jasmine.html).

Trying to do <script src="/templates/xyz.template"></script> does NOT work,
since in Backbone Views we do _.template($('#some_id').html()), and .html returns undefined.
THIS CANNOT BE CHANGED. there is no way to access the content of a script with src=xxx.
(look over in stack overflow if you dont belive me).

Then we created another base View class that inspects a templates property, and
does ajax $.get of the scripts, and populates the tag with the content, so that
_.template($('#some_id')) does work.

BUT NOW we need to wait for the templates to be fully loaded before rendering anything !